requirements.txt
youtube-transcript-api
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
python-dotenv
faiss-cpu
yt-dlp
sentence-transformers
psycopg[binary,pool]
langchain-postgres
psycopg2-binary
